Fix the incorrect byte write out totals
The value being returned by the `write_data` function was the number of
bytes read/written in the current loop iteration, not the total number
of bytes written across all the loop iterations in this function call.

Fixed by introducing a new total variable which correctly tracks the
number of bytes written across all iterations - the previous variable,
`size`, has been left in place, as it is used for other purposes.